Repeat Option
Added repeat option to announce.

If repeat is set to true, when the event reaches its 5 minute mark, it duplicates the event for a week later with the same parameters but different ID.

If repeat option is not entered, it is set to False.

Every 30 minutes, events that have already passed are deleted.